creating a new `xtdb.segment` package, ISegment -> Segment
