+++ core/src/main/kotlin/xtdb/segment/MergeTask.kt
@@ -1,0 +1,3 @@
+package xtdb.segment
+
+class MergeTask(val pages: List<Segment.Page>, val path: ByteArray)
